magentic/07-robotgame/requirements.txt
python-dotenv==1.0.1
autogen-agentchat==0.4.2
autogen-ext[openai,azure]==0.4.2
ipython==8.26.0